+++ artifacts/cli/src/index.ts
@@ -5,6 +5,8 @@ import { deployCommand } from "./commands/deploy.js";
 import { sitesCommand } from "./commands/sites.js";
 import { tokensCommand } from "./commands/tokens.js";
 import { rollbackCommand } from "./commands/rollback.js";
+import { analyticsCommand } from "./commands/analytics.js";
+import { statusCommand } from "./commands/status.js";
 
 const program = new Command();
 
@@ -14,9 +16,11 @@ program
   .version("0.1.0");
 
 program.addCommand(loginCommand);
+program.addCommand(statusCommand);
 program.addCommand(deployCommand);
 program.addCommand(sitesCommand);
 program.addCommand(tokensCommand);
 program.addCommand(rollbackCommand);
+program.addCommand(analyticsCommand);
 
 program.parse(process.argv);